% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_path}
\alias{add_path}
\alias{gen_path}
\alias{add_gen_path}
\title{Create PATHs to add to MATLAB PATHs}
\usage{
add_path(path)

gen_path(path)

add_gen_path(path)
}
\arguments{
\item{path}{path to add}
}
\value{
A character vector
}
\description{
Create PATHs to add to MATLAB PATHs
}
\examples{
add_path("~/")
gen_path("~/")
gen_path("~/")
}
