% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matlab_script.R
\name{have_matlab}
\alias{have_matlab}
\title{Logical check if MALTAB is accessible}
\usage{
have_matlab()
}
\value{
Logical TRUE is MALTAB is accessible, FALSE if not
}
\description{
Uses \code{get_matlab} to check if MATLAB is accessible
or the option
\code{matlab.path} is set and returns logical
}
\examples{
have_matlab()
}

