% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmat_to_matlab_mat.R
\name{rmat_to_matlab_mat}
\alias{rmat_to_matlab_mat}
\title{Convert R matrix to matlab matrix}
\usage{
rmat_to_matlab_mat(x, matname = NULL, transpose = FALSE)
}
\arguments{
\item{x}{matrix of values}

\item{matname}{Object in matlab to be assigned}

\item{transpose}{Transpose the matrix}
}
\value{
Character scalar of matlab code
}
\description{
This function takes in an R matrix then turns it into 
a matrix in matlab
}
