% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_script.R
\name{run_matlab_script}
\alias{run_matlab_script}
\title{Run matlab script}
\usage{
run_matlab_script(
  fname,
  verbose = TRUE,
  desktop = FALSE,
  splash = FALSE,
  display = FALSE,
  jvm = TRUE,
  figure_windows = TRUE,
  wait = TRUE,
  single_thread = FALSE,
  ...
)
}
\arguments{
\item{fname}{Filename of matlab script (.m file)}

\item{verbose}{print diagnostic messages}

\item{desktop}{Should desktop be active for MATLAB?}

\item{splash}{Should splash be active for MATLAB?}

\item{display}{Should display be active for MATLAB?}

\item{jvm}{should JVM be be loaded?  If \code{FALSE}, then
\code{-nojvm}}

\item{figure_windows}{should figure windows be enabled.  If
not, \code{-noFigureWindows} will be called}

\item{wait}{Should R wait for the command to finish.  Both
passed to \code{\link{system}} and adds the \code{-wait} flag.}

\item{single_thread}{Should the flag \code{-singleCompThread} 
be executed to limit MATLAB to a single computational thread?}

\item{...}{Options passed to \code{\link{system}}}
}
\value{
Exit status of matlab code
}
\description{
This function runs a matlab script, and 
returns exit statuses
}
