% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_script.R
\name{rvec_to_matlabclist}
\alias{rvec_to_matlabclist}
\title{Convert R vector to matlab cell mat}
\usage{
rvec_to_matlabclist(x, matname = NULL)
}
\arguments{
\item{x}{Character vector of values}

\item{matname}{Object in matlab to be assigned}
}
\value{
Character scalar of matlab code
}
\description{
This function takes in an R vector then turns it into 
a cell list
}

