% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_script.R
\name{get_matlab}
\alias{get_matlab}
\title{Find matlab path}
\usage{
get_matlab(try_defaults = TRUE)
}
\arguments{
\item{try_defaults}{(logical) If \code{matlab} is not found from 
\code{Sys.which}, and \code{matlab.path} not found, then try some 
default PATHs for Linux and OS X.}
}
\value{
Character of command for matlab
}
\description{
This tries to find matlab's path using a system which
command, and then, if not found, looks at \code{getOption("matlab.path")}.  If not path is found, it fails.
}

