% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_script.R
\name{have_matlab}
\alias{have_matlab}
\title{Logical check if MATLAB is accessible}
\usage{
have_matlab()
}
\value{
Logical \code{TRUE} is MATLAB is accessible, \code{FALSE} if not
}
\description{
Uses \code{\link{get_matlab}} to check if 
MATLAB's path accessible
}
\examples{
have_matlab()
}
