% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec_to_matlabcell.R
\name{rvec_to_matlabcell}
\alias{rvec_to_matlabcell}
\title{Convert R vector to matlab cell}
\usage{
rvec_to_matlabcell(x, sep = ";", matname = NULL, transpose = FALSE)
}
\arguments{
\item{x}{Character vector of values}

\item{sep}{separator to use to separate values. Defaults to ";"
argument}

\item{matname}{Object in matlab to be assigned}

\item{transpose}{Transpose the cell}
}
\value{
Character scalar of matlab code
}
\description{
This function takes in an R vector then turns it into 
a cell
}
